+++ html2asketch/helpers/background.js
@@ -1,0 +1,102 @@
+// Parses the background-image. The structure is as follows:
+// (Supports images and gradients)
+// ---
+// <background-image> = <bg-image> [ , <bg-image> ]*
+// <bg-image> = <image> | none
+// <image> = <url> | <image-list> | <element-reference> | <image-combination> | <gradient>
+// ---
+// Source: https://www.w3.org/TR/css-backgrounds-3/#the-background-image
+// ---
+// These functions should be pure to make it easy
+// to write test cases in the future.
+const parseBackgroundImage = value => {
+  if (value === 'none') {
+    return;
+  }
+
+  const urlMatches = value.match(/^url\("(.+)"\)$/i);
+  const linearGradientMatches = value.match(/^linear-gradient\((.+)\)$/i);
+
+  if (urlMatches && urlMatches.length === 2) {
+    // Image
+    return {
+      type: 'Image',
+      value: urlMatches[1]
+    };
+  } else if (linearGradientMatches && linearGradientMatches.length === 2) {
+    // Linear gradient
+    const linearGradientConfig = parseLinearGradient(linearGradientMatches[1]);
+
+    if (linearGradientConfig) {
+      return {
+        type: 'LinearGradient',
+        value: linearGradientConfig
+      };
+    }
+  }
+};
+
+// Parser for a linear gradient:
+// ---
+// <linear-gradient> = linear-gradient(
+//   [ [ <angle> | to <side-or-corner> ] ,]?
+//   <color-stop>[, <color-stop>]+
+// )
+//
+// <side-or-corner> = [left | right] || [top | bottom]
+// ---
+// Source: https://www.w3.org/TR/css3-images/#linear-gradients
+// ---
+// Example: "to top, rgba(67, 90, 111, 0.04), white"
+const parseLinearGradient = value => {
+  const parts = [];
+  let currentPart = [];
+  let i = 0;
+  let skipComma = false;
+
+  // There can be commas in colors, carefully break apart the value
+  while (i < value.length) {
+    const char = value.substr(i, 1);
+
+    if (char === '(') {
+      skipComma = true;
+    } else if (char === ')') {
+      skipComma = false;
+    }
+
+    if (char === ',' && !skipComma) {
+      parts.push(currentPart.join('').trim());
+      currentPart = [];
+    } else {
+      currentPart.push(char);
+    }
+
+    if (i === value.length - 1) {
+      parts.push(currentPart.join('').trim());
+    }
+    i++;
+  }
+
+  if (parts.length === 2) {
+    // Assume 2 color stops
+    return {
+      angle: '180deg',
+      stops: [parts[0], parts[1]]
+    };
+  } else if (parts.length > 2) {
+    // angle + n stops
+    const [angle, ...stops] = parts;
+
+    return {
+      angle,
+      stops
+    };
+  }
+
+  // Syntax is wrong
+  return null;
+};
+
+export {
+  parseBackgroundImage
+};
